Add EmptyState for source edit modal
