+++ src/components/SourceEditForm/parser/application.js
@@ -2,8 +2,11 @@ import React from 'react';
 import componentTypes from '@data-driven-forms/react-form-renderer/component-types';
 import validatorTypes from '@data-driven-forms/react-form-renderer/validator-types';
 import { FormattedMessage } from 'react-intl';
+import PlusCircleIcon from '@patternfly/react-icons/dist/esm/icons/plus-circle-icon';
+
 import { authenticationFields } from './authentication';
 import EditAlert from './EditAlert';
+import ResourcesEmptyState from '../../SourceDetail/ResourcesEmptyState';
 
 export const APP_NAMES = {
   COST_MANAGAMENT: '/insights/platform/cost-management',
@@ -51,10 +54,26 @@ export const applicationsFields = (applications, sourceType, appTypes) => [
       ...applications.map((app) => {
         const appType = appTypes.find(({ id }) => id === app.application_type_id);
 
+        const fields = createOneAppFields(appType, sourceType, app);
+
+        if (fields.length === 1) {
+          fields.push({
+            component: 'description',
+            name: 'no-credentials',
+            Content: ResourcesEmptyState,
+            message: {
+              id: 'resourceTable.emptyStateDescription',
+              defaultMessage: '{applicationName} resources will be added here when created.',
+            },
+            applicationName: appType?.display_name,
+            Icon: PlusCircleIcon,
+          });
+        }
+
         return {
           name: appType?.id,
           title: appType?.display_name,
-          fields: createOneAppFields(appType, sourceType, app),
+          fields,
         };
       }),
     ],
